- do not use 64bit integer instead of size_t (can't be alloc'ed), or when the actual possible values are 32bit or lower only
git-svn-id: http://svn.php.net/repository/php/php-src/trunk@315128 c90b9560-bf6c-de11-be94-00142212c4b1

+++ sapi/cli/php_http_parser.c
@@ -323,12 +323,12 @@ size_t php_http_parser_execute (php_http_parser *parser,
 {
   char c, ch;
   const char *p = data, *pe;
-  int64_t to_read;
+  size_t to_read;
 
   enum state state = (enum state) parser->state;
   enum header_states header_state = (enum header_states) parser->header_state;
-  uint64_t index = parser->index;
-  uint64_t nread = parser->nread;
+  uint32_t index = parser->index;
+  uint32_t nread = parser->nread;
 
   /* technically we could combine all of these (except for url_mark) into one
      variable, saving stack space, but it seems more clear to have them
@@ -1425,7 +1425,7 @@ size_t php_http_parser_execute (php_http_parser *parser,
       }
 
       case s_body_identity:
-        to_read = MIN(pe - p, (int64_t)parser->content_length);
+        to_read = MIN(pe - p, (size_t)parser->content_length);
         if (to_read > 0) {
           if (settings->on_body) settings->on_body(parser, p, to_read);
           p += to_read - 1;
@@ -1510,7 +1510,7 @@ size_t php_http_parser_execute (php_http_parser *parser,
       {
         assert(parser->flags & F_CHUNKED);
 
-        to_read = MIN(pe - p, (int64_t)(parser->content_length));
+        to_read = MIN(pe - p, (size_t)(parser->content_length));
 
         if (to_read > 0) {
           if (settings->on_body) settings->on_body(parser, p, to_read);
